@@ -1,7 +1,7 @@
 package com.zhanghaochen.smalldemos
 
 import android.content.Intent
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.View
 import com.zhanghaochen.smalldemos.demos.*
