@@ -2,7 +2,7 @@ package com.zhanghaochen.smalldemos.framework;
 
 import android.os.Handler;
 import android.os.Message;
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 
 import java.lang.ref.WeakReference;
 
