@@ -9,7 +9,7 @@ package com.zhanghaochen.smalldemos.framework.widgets.chartview;
 
 import android.content.Context;
 import android.graphics.Paint;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.util.AttributeSet;
 import android.view.View;
 
