@@ -36,7 +36,7 @@
  */
 package com.zhanghaochen.smalldemos.function.indicator.support;
 
-import android.support.v4.view.ViewPager;
+import androidx.viewpager.widget.ViewPager;
 import android.util.SparseArray;
 import android.util.SparseBooleanArray;
 
