@@ -37,7 +37,7 @@
 package com.zhanghaochen.smalldemos.function.recyclerviews;
 
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
