@@ -39,7 +39,7 @@ package com.zhanghaochen.smalldemos.function.recyclerviews.itemdecoration;
 import android.graphics.Canvas;
 import android.graphics.Paint;
 import android.graphics.Rect;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 
 /**
